requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
streamlit==1.31.1
python-multipart==0.0.9
python-dotenv==1.0.1
requests==2.31.0
pillow==10.2.0
python-jose==3.3.0
pydantic==2.6.1
aiohttp>=3.8.0
yt-dlp>=2023.12.30
beautifulsoup4>=4.12.0
